--- src/platform_impl/apple/mod.rs
@@ -1,6 +1,0 @@
-#[cfg(target_os = "macos")]
-mod appkit;
-
-#[allow(unused_imports)]
-#[cfg(target_os = "macos")]
-pub use self::appkit::*;
