+++ src/platform_impl/mod.rs
@@ -1,7 +1,5 @@
 #[cfg(android_platform)]
 mod android;
-#[cfg(macos_platform)]
-mod apple;
 #[cfg(any(x11_platform, wayland_platform))]
 mod linux;
 #[cfg(orbital_platform)]
@@ -11,13 +9,13 @@ mod web;
 #[cfg(windows_platform)]
 mod windows;
 
+#[cfg(macos_platform)]
+pub(crate) use winit_appkit as platform;
 #[cfg(ios_platform)]
 pub(crate) use winit_uikit as platform;
 
 #[cfg(android_platform)]
 use self::android as platform;
-#[cfg(macos_platform)]
-use self::apple as platform;
 #[cfg(any(x11_platform, wayland_platform))]
 use self::linux as platform;
 #[cfg(orbital_platform)]
