+++ winit-appkit/src/util.rs
@@ -2,10 +2,15 @@ use objc2_core_graphics::CGError;
 use tracing::trace;
 use winit_core::error::OsError;
 
+macro_rules! os_error {
+    ($error:expr) => {{
+        winit_core::error::OsError::new(line!(), file!(), $error)
+    }};
+}
+
 macro_rules! trace_scope {
     ($s:literal) => {
-        let _crate =
-            $crate::platform_impl::platform::appkit::util::TraceGuard::new(module_path!(), $s);
+        let _crate = $crate::util::TraceGuard::new(module_path!(), $s);
     };
 }
